unskip Failing test: Chrome UI Functional Tests.test/functional/apps/dashboard/group4/dashboard_empty·ts (#149889)
Fixes https://github.com/elastic/kibana/issues/149256

Flaky test runner
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/1824

Fix in https://github.com/elastic/kibana/pull/149582. `await
kibanaServer.importExport.unload(kbnDirectory);` is getting the `error
ECONNRESET`. Underlining cause determined to be @pgayvallet "Yeah, so
apparently it's related to the fact that
kibanaServer.importExport.unload calls bulkDelete with the summary from
the import file, which is not a valid SO. Not sure why it doesn't
constantly crashes on master though, as my changes in
https://github.com/elastic/kibana/pull/149582 makes it quite obvious"