[@kbn/handlebars] Simplify the way upstream changes are discovered (#150024)
Previously a custom diff of the files inside the upstream `spec`
directory was kept up-to-date in this package `.patches` directory. This
process was very tedious and wasn't providing much value.

In this commit I've simplified the process tremendously and simply rely on
checking if there are any new commits upstream and then allow the
developer to manually check for relevant changes. This is something they
needed to do with the old system regardless. Here the code is just much
simpler.

---------

Co-authored-by: Aleh Zasypkin <aleh.zasypkin@gmail.com>